Split into job and session modules
